3ju0gk-atomic-inventory-management-with-row-level-locking/requirements.txt
django==4.2.16
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-cors-headers